pyproject.toml
[project]
name = "virtualtree"
version = "0.1.0"
description = "Virtual Company - Multi-Agent AI Orchestration System"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0",
    "pyyaml>=6.0",
    "textual>=0.50",
    "watchfiles>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
vtree-tui = "src.tui.main:main"
